Rename submodules under `cupyx.scipy.sparse`
