+++ cupyx/scipy/sparse/_util.py
